move some functions back
